requirements.txt
# ============================================================
# Dependencias de Mikalia Bot
# ============================================================

# --- Core ---
anthropic>=0.42.0          # SDK oficial de Anthropic (Claude API)
gitpython>=3.1.0           # Operaciones Git desde Python
python-dotenv>=1.0.0       # Carga variables de .env
pyyaml>=6.0                # Parseo de config.yaml y front matter
croniter>=2.0.0            # Cron expression parsing para scheduler

# --- Semantic Memory ---
numpy>=2.0.0               # Operaciones vectoriales para embeddings
onnxruntime>=1.18.0        # Inferencia de modelo ONNX (ligero, sin PyTorch)
tokenizers>=0.19.0         # Tokenizacion rapida (HuggingFace)

# --- Document Analysis (F2) ---
pdfplumber>=0.10.0         # Extracción de texto de PDFs
python-docx>=1.1.0         # Lectura de archivos .docx

# --- GitHub App Auth ---
PyJWT[crypto]>=2.8.0       # JSON Web Tokens para GitHub App
cryptography>=42.0.0       # Firma RSA para JWT
requests>=2.31.0           # HTTP requests a GitHub API

# --- API Server ---
fastapi>=0.115.0           # Servidor HTTP (endpoints, webhooks)
uvicorn[standard]>=0.30.0  # ASGI server para FastAPI
python-multipart>=0.0.9    # Form parsing para webhooks Twilio

# --- Browser Control ---
playwright>=1.40.0         # Browser automation (Chromium headless)

# --- Voice ---
edge-tts>=6.1.0            # Text-to-Speech (Microsoft Neural, gratis)
faster-whisper>=1.0.0      # Speech-to-Text (Whisper ONNX, local)

# --- Data & Reports ---
matplotlib>=3.8.0          # Graficas y visualizaciones de datos
fpdf2>=2.7.0               # Generacion de PDFs (puro Python, ligera)

# --- Telegram ---
python-telegram-bot>=21.0  # SDK de Telegram Bot API

# --- UX & Logging ---
rich>=13.0.0               # Output bonito en terminal (colores, spinners, tablas)
click>=8.1.0               # CLI framework (alternativa más moderna a argparse)

# --- Testing ---
pytest>=8.0.0              # Framework de testing
pytest-asyncio>=0.23.0     # Tests async
pytest-mock>=3.12.0        # Mocking para tests

# --- Calidad de código ---
ruff>=0.2.0                # Linter + formatter (reemplaza flake8+black+isort)
mypy>=1.8.0                # Type checking estático
